Got initial system test running
The system test currently just creates a namespace and tears it down.
Most of this code should be reusable by anyone wanting to write a system
(integration? not sure the best terminology to use here, working on it!)
test for their CRDs - though this implementation requires `client-go`.

Using `dep init` + `dep ensure` to manage dependencies.

Next step will be to to get the controller up and running with a simple
integration test.