Get the initial controller running
This controller is pretty much a hacked up version of
https://github.com/kubernetes/sample-controller The resources are very
similar, if you create a Cat, a deployment of an nginx service with the
same name as the Cat gets created.

The Cat's status is not yet updated.

Also added the config yamls that I think I need to deploy with `ko`
but I haven't tried it yet!