Fix controller
When I tried to run the end to end tests, I realized that the controller
wasn't starting. This was first b/c the path to the repo wasn't correct
so `ko` wasn't even building and deploying the controller.

The next problem was the controller pod waited for the config-logging
config-map to be created (which it doesnt actually need).

Lastly the controller didn't have permission to access apps.deployments.

These issues all came from naively copy-pasting the config from the
build-pipline repo lol