Focus README on test examples
In the end we have 2 types of end to end tests:
- client-go based
- kubebuidler based

And we have the kubebuilder integration tests.

Maybe one day (or not) it would be nice to add the client-go integration
tests (with test doubles) and some actual unit tests (but I'd probably
need to make the funcitonality more complicated to be able to do that).